requirements.txt
crewai_tools==0.2.6
exa_py==1.0.9
fakeredis==2.23.1
html2text==2020.1.16
langchain_community==0.0.38
langchain_core==0.1.52
langchain_groq==0.1.4
langchain_openai==0.1.7
openai==1.30.1
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
Requests==2.31.0
retry==0.9.2
streamlit==1.29.0
streamlit_tags==1.2.8
tiktoken==0.7.0
